Refactor game logic to have index number cycle through the array without removing
This branch is to refactor the code to avoid removing items from the questions array, which meant that the array had to be repopulated at the end of every game. Now, the array will remain at a fixed number, and will be shuffled on hitting play. The way to cycle through the array is just by incrementing the index, 'QNumber'.

This also makes consequential changes to features such as the progress bar that use a Questions.count logic.

Finally, the system sounds are working for the custom sound but not the device sounds.